リファクタリング途中

+++ Project_K/Util/SelectUI.cpp
@@ -268,7 +268,7 @@ void SelectUI::DrawChar()
 
 	// 1Pが選択してるキャラ
 	my::MyDrawRectRotaGraph(
-		400, m_player1.m_posY,//プレイヤーの位置
+		400, static_cast<int>(m_player1.m_posY),//プレイヤーの位置
 		m_player1.m_imgWidth * m_indexWidth1, m_player1.m_imgHight * m_indexHight1,// 画像の右上
 		m_player1.m_imgWidth, m_player1.m_imgHight,// 画像の切り取り
 		m_player1.m_imgSize,// 画像の拡大率
@@ -279,7 +279,7 @@ void SelectUI::DrawChar()
 
 	// 2Pが選択してるキャラ
 	my::MyDrawRectRotaGraph(
-		Game::kScreenWidth - 400, m_player2.m_posY,//プレイヤーの位置
+		Game::kScreenWidth - 400, static_cast<int>(m_player2.m_posY),//プレイヤーの位置
 		m_player2.m_imgWidth * m_indexWidth2, m_player2.m_imgHight * m_indexHight2,// 画像の右上
 		m_player2.m_imgWidth, m_player2.m_imgHight,// 画像の切り取り
 		m_player2.m_imgSize,// 画像の拡大率
@@ -295,8 +295,6 @@ void SelectUI::DrawLetter()
 	DrawStringToHandle((Game::kScreenWidth -
 		GetDrawStringWidthToHandle(kSelectText, 16, kSelectFont)) / 2,
 		100 - 35, kSelectText, 0x800000, kSelectFont);
-
-	//int a = GetDrawStringWidthToHandle(kSelectText, 12, m_player1.m_font) / 2;
 
 	// 1Pの名前表示
 	DrawStringToHandle(400,
